// src/models/Enum.ts
//ENUMS
export enum Movies  {
    MAX_NUMBER_OF_TICKETS = 100,
    NUMBER_OF_SCREENS = 5,
    NUMBER_OF_WINDOWS =     4
}
export enum CanteenInventory  {
    NUMBER_OF_WATER_BOTTLES = 500,
    NUMBER_OF_POPCORNS = 500,
    NUMBER_OF_SODAS = 200
}
export enum RecordType  {
    MOVIE = "movie",
    THEATRE = "theatre",
    TICKET = "ticket" 
}
export enum Status  {
    ACTIVE = "Active",
    INACTIVE = "Inactive"
}